Refactor all the crud and data logic to app.js
