+++ index.html
@@ -12,18 +12,17 @@
       integrity="sha384-HzLeBuhoNPvSl5KYnjx0BT+WB0QEEqLprO+NBkkk5gbc67FTaL7XIGa2w1L0Xbgc"
       crossorigin="anonymous"
     />
-    <!-- <script type="module" src="./index.js"></script> -->
   </head>
 
   <body>
     <h1 class="title">Yet Another Todo</h1>
     <div class="container">
-      <todo-list></todo-list>
+      <todo-app></todo-app>
     </div>
   </body>
-
   
   <script type="module" src="./src/components/addForm/index.js"></script>
   <script type="module" src="./src/components/todoItem/index.js"></script>
   <script type="module" src="./src/components/todoList/index.js"></script>
+  <script type="module" src="./src/app.js"></script>
 </html>
